+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   - com.dwarfeng.familyhelper.life.stack.service.ActivityOperateService。
   - com.dwarfeng.familyhelper.life.stack.service.ActivityCoverOperateService。
   - com.dwarfeng.familyhelper.life.stack.service.ActivityFileOperateService。
+  - com.dwarfeng.familyhelper.life.stack.service.ActivityParticipantOperateService。
 
 - 新增实体及其维护服务，单元测试通过。
   - com.dwarfeng.familyhelper.life.stack.bean.entity.Activity。